requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.17.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
bottle==0.12.25
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-pagination==0.12.10
fastapi-socketio==0.0.10
frozenlist==1.4.0
h11==0.14.0
html2text==2020.1.16
html5lib==1.1
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.3
markdown2==2.4.10
MarkupSafe==2.1.3
multidict==6.0.4
nicegui==1.3.16
numpy==1.26.0
orjson==3.9.7
packaging==23.2
Pillow==10.0.1
proxy-tools==0.1.0
pscript==0.7.7
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Security==10.0
pyobjc-framework-WebKit==10.0
pypika-tortoise==0.1.6
PySimpleGUI==4.60.5
python-dotenv==1.0.0
python-engineio==4.7.1
python-imap==1.0.0
python-multipart==0.0.6
python-socketio==5.9.0
pytz==2023.3.post1
pywebview==4.3.3
PyYAML==6.0.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tk==0.1.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.22.0
uvloop==0.17.0
vbuild==0.8.2
watchfiles==0.20.0
webencodings==0.5.1
websockets==11.0.3
wsproto==1.2.0
wxPython==4.2.1
yarl==1.9.2
